revert TS changes
